+++ crates/aiken-lang/src/gen_uplc.rs
@@ -3759,17 +3759,15 @@ impl<'a> CodeGenerator<'a> {
 
                     interner.program(&mut program);
 
-                    let eval_program: Program<NamedDeBruijn> =
-                        program.clean_up_no_inlines().try_into().unwrap();
+                    let cleaned_program = program.clean_up_no_inlines();
 
-                    Some(
-                        eval_program
-                            .eval(ExBudget::max())
-                            .result()
-                            .unwrap_or_else(|e| panic!("Failed to evaluate constant: {e:#?}"))
-                            .try_into()
-                            .unwrap(),
-                    )
+                    let eval_program: Program<NamedDeBruijn> =
+                        cleaned_program.clone().try_into().unwrap();
+
+                    Some(match eval_program.eval(ExBudget::max()).result() {
+                        Ok(term) => term.try_into().unwrap(),
+                        Err(_) => cleaned_program.term,
+                    })
                 }
                 ValueConstructorVariant::ModuleFn {
                     name: func_name,